requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
cssselect==1.2.0
cssutils==2.11.1
Django==5.1.1
djangorestframework==3.15.2
idna==3.8
lxml==5.3.0
more-itertools==10.4.0
numpy==2.1.1
pandas==2.2.2
premailer==3.10.0
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
XlsxWriter==3.2.0
